pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "patentbench"
version = "0.1.0"
description = "The first reproducible benchmark for patent prosecution AI"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "Salt Holdings LLC", email = "engineering@saltholdings.com"},
]
keywords = ["patent", "benchmark", "ai", "prosecution", "legal", "evaluation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "click>=8.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "tqdm>=4.65",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.5",
]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.25"]
google = ["google-generativeai>=0.5"]
all = [
    "patentbench[openai]",
    "patentbench[anthropic]",
    "patentbench[google]",
]

[project.scripts]
patentbench = "scripts.run_benchmark:main"

[project.urls]
Homepage = "https://github.com/salt-holdings/patentbench"
Documentation = "https://abigail.app/patentbench"
Repository = "https://github.com/salt-holdings/patentbench"
Issues = "https://github.com/salt-holdings/patentbench/issues"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
